refactor: split HCL symbol construction into extract/hcl.rs (ADR 0028)
Move hcl_block_type, hcl_string_lit_text, hcl_block_name, and
build_hcl_locals_symbols out of extract/mod.rs (1003 lines, over the
1000-line warn threshold) into a new sibling module extract/hcl.rs,
the same split references.rs got for the same reason. Add the
rationale doc comments the original commit omitted (review finding).

Pure move: no logic changes, only visibility (pub(super) for the
three functions called from mod.rs) and super:: path adjustments.
